+++ src/quantity/arithmetic/scalar.rs
@@ -1,6 +1,6 @@
 use core::ops::{Div, Mul};
 
-use eunomia::FloatElement;
+use eunomia::{Complex, FloatElement, RealField};
 
 use crate::quantity::Quantity;
 
@@ -25,5 +25,29 @@ where
     #[inline]
     fn div(self, rhs: T) -> Self::Output {
         Self::from_base(self.value / rhs)
+    }
+}
+
+impl<T, D> Mul<T> for Quantity<Complex<T>, D>
+where
+    T: RealField,
+{
+    type Output = Self;
+
+    #[inline]
+    fn mul(self, rhs: T) -> Self::Output {
+        Self::from_base(self.value.scale(rhs))
+    }
+}
+
+impl<T, D> Div<T> for Quantity<Complex<T>, D>
+where
+    T: RealField,
+{
+    type Output = Self;
+
+    #[inline]
+    fn div(self, rhs: T) -> Self::Output {
+        Self::from_base(self.value.scale(rhs.recip()))
     }
 }