+++ src/systems/si/units/mod.rs
@@ -9,9 +9,10 @@ pub use derived::{
     Coulomb, CubicMeter, CubicMeterPerSecond, Farad, Gray, GrayPerSecond, Hertz, Joule,
     JoulePerCubicMeter, JoulePerKelvin, JoulePerKilogramKelvin, JoulePerMole, JoulePerMoleKelvin,
     JoulePerSquareMeter, KilogramPerCubicMeter, KilogramPerCubicMeterSecond, MeterPerSecond,
-    Newton, NewtonPerMeter, Pascal, PascalSecond, PerCubicMeter, PerKelvin, PerMeter, PerSecond,
-    PerSquareKelvin, Rayl, Siemens, SquareMeter, SquareMeterPerKilogram, SquareMeterPerSecond,
-    Volt, Watt, WattPerCubicMeter, WattPerKilogram, WattPerMeterKelvin, WattPerSquareMeter,
+    Newton, NewtonPerMeter, Ohm, Pascal, PascalSecond, PerCubicMeter, PerKelvin, PerMeter,
+    PerSecond, PerSquareKelvin, Rayl, Siemens, SquareMeter, SquareMeterPerKilogram,
+    SquareMeterPerSecond, Volt, Watt, WattPerCubicMeter, WattPerKilogram, WattPerMeterKelvin,
+    WattPerSquareMeter,
 };
 pub use scaled::{
     Centimeter, CubicMillimeter, ElectronVolt, Gram, GramPerCubicCentimeter, JoulePerMilliliter,